update mhc to 0426 version

+++ csrc/moe/hc_pre_inv_rms/op_kernel/hc_pre_inv_rms_full_load_regbase.h
@@ -1,10 +1,10 @@
 /**
- * This program is free software, you can redistribute it and/or modify.
- * Copyright (c) 2025 Huawei Technologies Co., Ltd.
- * This file is a part of the CANN Open Software.
- * Licensed under CANN Open Software License Agreement Version 2.0 (the "License").
+ * Copyright (c) 2026 Huawei Technologies Co., Ltd.
+ * This program is free software, you can redistribute it and/or modify it under the terms and conditions of
+ * CANN Open Software License Agreement Version 2.0 (the "License").
  * Please refer to the License for details. You may not use this file except in compliance with the License.
- * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
+ * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED,
+ * INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
  * See LICENSE in the root of the software repository for the full text of the License.
  */
 
@@ -129,11 +129,10 @@ __aicore__ inline void HcPreInvRmsFullLoadRegbase<T>::Process()
 template <typename T>
 __aicore__ inline void HcPreInvRmsFullLoadRegbase<T>::CopyIn(uint64_t idx, uint64_t curUbFactorA)
 {
-    
     LocalTensor<T> xLocal = inQueueX.AllocTensor<T>();
 
     DataCopyPadExtParams<T> dataCopyPadParams{false, 0, 0, 0};
-    int32_t xGmStartAddr = idx * R * ubFactorA;
+    int64_t xGmStartAddr = idx * R * ubFactorA;
     DataCopyExtParams dataCopyParams{
         static_cast<uint16_t>(curUbFactorA), static_cast<uint32_t>(R * sizeof(T)), 0, 0, 0};
     DataCopyPad(xLocal, xGm[xGmStartAddr], dataCopyParams, dataCopyPadParams);
@@ -302,4 +301,3 @@ __aicore__ inline void HcPreInvRmsFullLoadRegbase<T>::CopyOut(uint64_t idx, uint
 
 } // namespace HcPreInvRmsRegbase
 #endif // ASCENDC_HC_PRE_INV_RMS_FULL_LOAD_REGBASE_H_
-
